Fixed math with total price, Moved Network Modem from Kitchen to living room and tested to try repeate phone bug could not replicate

+++ BobsBuildingSupplys/Mod_Varivbles.vb
@@ -8,14 +8,14 @@
     Public cusTrade As Boolean = False
     Public cusAdressSame As Boolean = False
     Public Kitchen(5) As Integer
-    Public kitchencost(5) As Integer
+    Public kitchencost(5) As Decimal
     Public BathRoom(2) As Integer
-    Public BathRoomCost(2) As Integer
+    Public BathRoomCost(2) As Decimal
     Public LivingRoom(5) As Integer
-    Public LivingRoomCost(5) As Integer
+    Public LivingRoomCost(5) As Decimal
     Public BedRoomTwo(5) As Integer
-    Public BedRoomTwoCost(5) As Integer
+    Public BedRoomTwoCost(5) As Decimal
     Public BedRoomOne(5) As Integer
-    Public BedRoomOneCost(5) As Integer
+    Public BedRoomOneCost(5) As Decimal
     Public totalvalue(5) As Integer ' 0 = total,  1 = kitchen,  2 = living room,  3= bedroom1, 4 = bedroom2, 5= bathroom
 End Module
